Completely removed NullScope, refactored ruby_code method and ResourceReference type_name
NullScope was removed because it could be a source
for future bugs.
Removed ruby_code= method and ruby_code returns an
array now.
ResourceReference had type ivar, but it's name was
suggesting it's an object, not a string, so I
renamed it to type_name.

There are also some minor updates like updated
tests and conversion from ivars to let in tests.