Maint: Clear global state between tests
Previously, the tests would fail when invoked when random seed 3366,
because of global state associated with the Basetest face application,
in particular, option_parser_commands, which contained the list of valid
commands from the previous test.

This commit adds a method to clear application global state. Note that
the state is not associated with Puppet::Application, but the subclass
that the test defines, which is why test_helper doesn't call it.