Maint: Clear memory instances
Previously, the test would fail when invoked with random seed 3366,
because the `instances` variable associated with the memory terminus was
not cleared between tests. As a result, the test that expected not to
find any data, would fail.

Note the test_helper only restores the indirections' cache_class and
terminus_class. It does not save/restore the indirections' termini or
any data maintained by the termini.

This commit ensures the memory termini's instances variable is cleared
between runs.