Maint: Don't overwrite outer sum variable
Previously, these tests failed with random seed 3366, because one of the
tests declared a block whose inner variable overwrote the outer variable
with the same name. So a subsequent test that expected sum to be :md5
would fail.

This commit just changes the inner block variable to not collide with
the outer one.