Maint: Ensure manages_password_age feature is set
Previously, the test would fail with random seed 31725 on systems where
the manages_password_age feature is not enabled by default. This
occurred because one of the tests was declaring the useradd provider
supported the manages_password_age feature, which seems to have global
scope. Rather than try to debug that now, I've updated the other tests
that rely on the manages_password_age feature being present to follow
the same pattern.