(#16144) Update puppet spec file for fedora 17 and systemd changes
This commit adds some conditional logic to puppet's spec file for fedora 17 and
greater. Specifically, fedora 17 has systemd as the default service, so the
spec will install systemd service files on fedora >= 17, and make sure systemd
reloads its daemons to notice the new puppet service. The same changes are
applied to the puppet-server subpackage. On removal, the systemd services
are removed.