(#3757) Rename agent_pidfile to agent_catalog_run_lockfile
Without this patch it is very confusing what the intended purpose of the
agent_pidfile setting is.  This confusion may lead to people thinking
this file is intended to be used my monitoring systems or service
management frameworks to determine the PID of a running puppet agent
process.  This thinking is incorrect, the `pidfile` setting is intended
for this purpose.

The `agent_pidfile` setting is intended to provide mutual exclusion of
configuration catalog runs.  This patch reduces confusion by more
closely matching the name and description to the intent.

In addition, the prose in the descriptions and comments have been
augmented to make it clear "running agent" refers to an agent performing
a configuration catalog run and not simply the running agent process.