(Maint) Improve error message when catalog run is in progress
When a catalog run is currently in progress, or the `puppet agent
--configprint agent_catalog_run_lockfile` is stale, the error message
looks like this:

    Run of Puppet configuration client already in progress; skipping

This is a problem because if the system is deadlocked from a stale file,
it's not clear what file needs to be removed.  This patch improves the
error message to be:

    Run of Puppet configuration client already in progress; skipping \
      (/var/lib/puppet/state/agent_catalog_run.lock exists)

A new method has been added to make it easy to print out the file being
referenced by the `running?` method.