(Maint) Make puppet module install --confdir=foo puppetlabs-stdlib work
Without this patch I've been annoyed for quite some time that a simple
`puppet module --confdir=/tmp/puppet install puppetlabs-stdlib` errors
out with the message:

    Err: 'module' has no 'default' action.  See `puppet help module`.

The root cause of this issue is that we overlooked the situation where a
long-form argument such as `--confdir` contains the mandatory argument
in its own string.  We assume the mandatory argument is the next item in
the argument vector, but it may not be the next item if the user
specified the mandatory argument as `--confdir=/tmp/foo`.

This patch fixes the problem by not skipping the next item in this
scenario.