Update puppet.spec.erb to correct rpmlint issues
This commit updates the puppet rpm specfile for several warnings and errors
issued by rpmlint against our packages. These include incorrect perms on
scripts and self-obsoleting package. This also replaces Requires(stage) with a
simple requires with a comment explaining which stage it is required for and
adds cases for systemd requires.