(#7316) Load applications from the modulepath
Without this patch Puppet does not search the modulepath lib
subdirectories for application subcommands.  This is a problem because
face applications that aren't affected by #3947 should work.

This patch fixes the problem by searching through the modules along the
modulepath to find available subcommands in addition to those available
along the Ruby $LOAD_PATH and RubyGems $GEM_PATH.

Both legacy application files and modern face applications are loaded
from the modulepath with this patch.

This patch does _not_ add the module's `lib/` subdirectory to the
$LOAD_PATH, so issue #3947 will still affect face applications that
attempt to use `require` with library files inside the module.