Update debian ruby dependencies to include 1.9
The previous set of debian dependencies were all ruby 1.8 specific. Puppet 3.x
has full ruby 1.9 support, so this commit adds the option of ruby 1.9 for all
of Puppet's dependencies. It moves the sitelibdir to be vendordir, which is the
current debian standard for ruby projects with ruby support for both 1.8 and
1.9. It also updates all of the dependencies to be just ruby instead of ruby1.8
| ruby 1.9.1 and adds 1.9 equivalents for all ruby libraries. This commit also
drops rails, activerecord, and mongrel and adds the suggestion of
puppetdb-terminus instead.