(Maint) Fix spec failures with Puppet::Face[]
Without this patch the describe blocks are being executed before the
spec helper before hooks.  This is problem because the
Puppet::Face.find method now searches the modulepath.  The modulepath
setting is not yet initialized when the subject of the describe block is
evaluated.

This patch fixes the problem by explicitly setting the subject inside
the example groups.  Doing so delays evaluation of the find method until
after settings have initialized.

I suppose all of this could be summed up as, "Load faces after settings
have been initialized by the spec helper, not before."

Paired-with: Andrew Parker <andy@puppetlabs.com>