(maint) Fix Failed to load face errors
Without this patch applied Telly will spew out a bunch `Error: Failed to
load face foobar` errors when `puppet help` is run.  These errors look
like:

    Error: Failed to load face help:
    no such file to load -- puppet/face/0.0.1/help
    Error: Failed to load face ca:
    no such file to load -- puppet/face/0.1.0/ca
    Error: Failed to load face catalog:
    no such file to load -- puppet/face/0.0.1/catalog
    Error: Failed to load face certificate:
    no such file to load -- puppet/face/0.0.1/certificate
    Error: Failed to load face certificate_request:
    no such file to load -- puppet/face/0.0.1/certificate_request
    Error: Failed to load face certificate_revocation_list:
    no such file to load -- puppet/face/0.0.1/certificate_revocation_list
    Error: Failed to load face config:
    no such file to load -- puppet/face/0.0.1/config
    ..etc.. for each face, but then see the regular help output.

The problem is that the regular expression changed in a previous patch
in my #7316 work, causing load errors resulting from checking for
version specific faces to pass through the exception handler.

This patch fixes the problem by tightening up the regular expression to
catching both version specific faces and "current" face load errors.

The `Puppet.debug` statement is also wrapped inside of a conditional so
that we don't log spam and we can accurately see if we're actually
loading multiple times or not.  The expectation is that `require` will
only load the file once and return false on subsequent calls.