(#15081) Make virtual resources work with create_resources
Without this patch the create_resources function is unable to create
virtual or exported resources.  This is a problem because the resources
created by create_resources are not a first class citizen with respect
to resources created from the Puppet DSL, or other means.

This patch fixes the problem by allowing the create resources function
to accept type names prefixed with @ or @@.

For example, the following puppet manifest works with this patch
applied.

    create_resources('@file', {'/etc/foo' => {'ensure'=>'present'}})
    realize(File['/etc/foo']