16553 Exclude portmap-wait from self.instances
Previous to this commit, running self.instances on Ubuntu 10.04 broke
due to a bug in the portmap-wait init script. The bug made it impossible
to retrieve the state of the service. This service doesn't really have a
status since it's only meant to run once on boot.

This commit adds the portmap-wait service to the exclude list when
running self.instances