Pass wrapped real scope to DataBinding
We were passing only facts into the DataBinding terminus, but Hiera needs to
be able to lookup variables on the scope. We now pass a scope, but wrap it
with Puppet::DataBinding::Variables to avoid exposing anything but variable
lookup.

This fixes the error in Puppet 3.0rc7 where variables like %{environment} were
not properly interpolated in hiera.conf (and likely elsewhere).