(#7316) Do not set default values for confdir and vardir in the master
Without this patch applied `puppet master` will explicitly set the value
for confdir and vardir inside of
Puppet::Application::Master#app_defaults.  This is a problem because
it calls a method that does not exist while trying to do so.

This patch fixes the problem by implementing the intended behavior
documented in #15337, and specifically in commit 38cba0a.  This intended
behavior is:

  1: If provided, use explicit puppet.conf in `--confdir`
  2: If root, use system puppet.conf
  3: Otherwise, use ~/.puppet/puppet.conf