(#16637) Fix default confdir and vardir with puppet master as normal user
Without this patch `puppet master` will select the wrong default confdir
and vardir values when running as a non root user.  This is a problem
because permission denied errors result trying to create the system
directories when running as a non-system user.

This patch fixes the problem by removing the explicit check for the
master run mode in Puppet::Util::RunMode#which_dir