#(16271) Implement restart action for launchd
Previously, the launchd service provider did not accommodate a restart
action. If a resource attempted the refresh the service, Puppet triggered
an alert but did not actually restart the service. This commit adds a
restart method that calls the stop and then start action on the service
(since launchd doesn't actually have a restart action). Puppet.debug
messages are also raised for debugging purposes.