(#15717) Acceptance tests for puppet kick
Previously, the acceptance test for #3172 only verified that `puppet
kick` could be given a target host to connect to, and would verify that
exit code 3 was returned. This exit code can mean connect failed, which
is what the test was attempting to verify.

However, exit code 3 can also mean `puppet kick` failed to verify the
SSL server's certificate, or auth.conf doesn't allow access to /run, or
something bad happened on the remote agent end -- in our case #15717.

This commit updates the acceptance test for #3172 so it's clear what
it's actually testing.

It also adds an acceptance test to verify that the master can kick an
agent, and that the agent reports the kick was successful. It also
verifies that running `puppet kick` generates a deprecation warning.

While researching these issues, I've discovered that `--ignoreschedules`
doesn't work and hasn't for sometime, see #5975. The option is also not
documented in the puppet kick documentation.

Also, the `--pluginsync` option doesn't do anything, since in Puppet 3,
pluginsync is enabled by default, and the ability to specify
`--no-pluginsync` doesn't make much sense in an agent/master
environment.