(#16676) Clarify file format restrictions around comments
Previously the help text at the top of the generated config files
did not describe what the file format ought to be. There is a
description in the documentation (guides/configuration.html),
but our inline docs should be consistent with that. This led
to confusion and bugs like #16676 and #16672.

Now we state clearly that whole-line comments are allowed but
not partial-line comments. Plus I got to use the word 'octothorpe'.