(#16757) Move rubygems require to application entry point
The previous commit removed the side-effect of loading rubygems as a
result of evaluating the `Puppet.features.rubygems?` feature. As a
result, the puppet script was the only place where we loaded rubygems,
breaking rack setups.

This commit moves the loading of rubygems into the
Puppet::Util::CommandLine, which is effectively puppet's entry point
for both command line and rack setsups.

In addition, `Bundler::Setup` doesn't always seem to exist, at least
not in all versions of bundler, so instead we're using `::Bundler` to
detect if it's loaded.

Paired-with: Andy Parker <andy@puppetlabs.com>