(#16757) Deprecate Puppet.features.rubygems?
Loading rubygems inside an library is wrong, it needs to be controlled
by the environment.

Previously, several puppet features relied on the side effect of
loading rubygems as a result of evaluating the feature. This is no
longer necessary now that rubygems loading is handled in the
Puppet::Util::CommandLine entry point.

This commit adds a deprecation warning if `Puppet.features.rubygems?`
is invoked, and reinstates the previous behavior of requiring rubygems
to maintain backwards compatibility.

Paired-with: Andy Parker <andy@puppetlabs.com>