(#16801) Remove mongrel references from init scripts
Built-in mongrel support has been removed in Puppet 3.x, so the puppetmaster
init scripts and defaults files should no longer mention them or try to use
them as options. This commit removes the code paths that reference mongrel and
--servertype. It also adds checks to both init scripts to check for multiple
ports or puppetmasters set. In those cases it prints a warning to the user
explaining that only one puppetmaster, on one port, will be started. It also
explains where the setting can be updated.