(#16801) Update suse server.init and remove mongrel refs
The binary used to start the puppetmaster process has changed in Puppet 3.0, as
has the pidfile location. Built-in mongrel support has also been removed. This
commit adds a warning if mongrel settings are detected and also updates calls
to startproc, checkproc, and killproc to use a pidfile to ensure the correct
process is used, and not the `puppet agent` process. It also updates the
pidfile location to match reality and adds master to the PUPPETMASTER_OPTS so
the correct application is started.