Update suse client.init for new pidfile, process name
The binary used to start the puppet agent process has changed in Puppet 3.0, as
has the pidfile location. This commit updates calls to startproc, checkproc,
and killproc to use a pidfile to ensure the correct process is used, and not
the `puppet master` or other process. It also updates the pidfile location to
match reality and adds agent to the PUPPET_OPTS so the correct application is
started.