(#16698) Do not set Node cache_terminus for storedconfigs
The hook on the storedconfigs setting would try and setup the terminuses
so that they were properly wired up to the store configs. One part of
this was to set the `cache_terminus` for the `Puppet::Node` to
`:store_configs`. However, up until commit
5a79d9abd96e73ff166527cdee69a30da8ab0f87 this was overridden by the
master application to be `:yaml`. That commit removed this so that an
ENC was able to be authoritative, but once it was removed the store
configs cache terminus was left on and interferred with the ENC.

This commit removes the cache for the `Puppet::Node` since there is no
need for it and all it does is interfere with fetching node data. The
`Puppet::Node::Facts` store configs code will actually create the entry
in the `host` table that is needed.