(Maint) Update config.ru to fix issue with vardir
Without this patch the config.ru example only contains the --confdir
argument.  This is a problem because when Puppet starts from config.ru,
it will automatically be switched by Passenger to the EUID of the owner
of the file.  Puppet will still try to write to the default vardir of
~/.puppet/ in this situation which may be a problem because $HOME might
be ~root.

This patch fixes the problem by adding the --vardir option to the
example with the default value of /var/lib/puppet