(#16581) Use native Win32 APIs to resolve SIDs in file provider
Previously, the file provider used the ADSI module to resolve SIDs, which
is very slow in domain environments.

This commit changes it to use the Puppet::Util::Windows::Security module
instead. Ideally, the `name2id` and `id2name` methods could be eliminated
and instead have the uid & gid methods delegate directly to the Security
module. But they can't, because the Security module is not require'd on
non-Windows platforms.