(#16581) Use win32-security gem to resolve SIDs
This commit adds methods for resolving Windows account names into binary
encoded SIDs and back. Account names may be specified with or without the
domain component, or as a SID string. So all of the following are
equivalent names for the LocalSystem account:

    NT AUTHORITY\SYSTEM
    SYSTEM
    S-1-5-18

There is a bug[1] in the `SID.string_to_sid` method in versions of
win32-security prior to 0.1.4, so this commit updates the Gemfile
accordingly.

[1] https://github.com/djberg96/win32-security/pull/1

Conflicts:

	spec/unit/util/windows/sid_spec.rb