(#16798) Update revocation by serial number documentation
Without this patch the documentation mis-represents how revocation by
certificate serial number occurs.  When revoking by serial number rather
than common name, the serial number must be represented as a hex-encoded
string with a prefix of '0x'.

This patch simply removes the mention of representing the serial number
as a decimal encoded string.