(#16798) Add more tests to cover revoke by serial number
Without this patch there is only one test covering the revoke by serial
number behavior.  This isn't a problem in and of itself, but I'd like to
see a bit more coverage around the common behavior of mixed-case and the
not-so-common behavior of extremely large serial numbers.

This patch adds spec tests for both behaviors.