(#17129) Refactor code for loading ssl files
Prior to this commit, the code for creating a model and loading it from
a file was duplicated in two places. Also, the
`renaming_files_with_uppercase` method was being called after the
existence tests, complicating the logic.

This commit moves the duplicate code to a single method. It also moves
the `renaming...` method earlier, so that we get into a sane state as
soon as possible.

Finally, it adds a deprecation warning for some crazy behavior in the
`rename_files_in_uppercase`. The `search` method may return an uppercase
name, since it just walks the directory. But if `find` is called on the
uppercase name, then the file gets lowercases, so that subsequent calls
to `search` return the lowercased name. See #17307.