(#15513) Clarify target attribute for cron
Without this patch applied the behavior of `target` in the cron type isn't made clear.
This is a problem because users report trying to use a file as the target instead of
the intended username, producing unexpected results.

This patch clarifies the behavior of `target` by improving the description docstring.