(#16581) Fix mis-stubbed test
Previously, the spec test failed because it was stubbing the
`sid_for_account` method that the scheduled_task provider no longer
called. This wasn't noticed on my system, because I had a local 'bob'
user, so that test would still "pass" incorrectly, but fail in CI.

This commit removes the `stubs` and adds `expects` for each test as
appropriate.