(#16791) Re-instate deprecated symbolize and symbolizehash!
The cherry-pick of commit a09df46f58187ad9ff0f599de47cd8c1e0a417eb to
remove the use of symbolize throughout the code base also removed the
symbolize method, as well as the symbolizehash! method. The cherry-pick
was needed so that the methods could be deprecated without puppet
issuing warnings from using them.

This commit puts the two methods back and also places a deprecation
warning in them.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch ticket/2.7.x/16791-cleanup-performance-backports
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   lib/puppet/util.rb
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	.irb-history
#	.rvm/
#	acceptance/tests/basic-datacenter.rb
#	debug-requests.patch
#	doc/
#	ext/packaging/
#	lib/puppet/network/http/handler.rb.orig
#	pa
#	required-home
#	scripting
#	t.yaml
#	time-exec.rb