(#17158) clarify documentation on behavior of `replace` toward symlinks in file type
Without this patch applied the behavior of `replace` toward symlinks in the file type isn't
made explicit. This is a problem because some other parameters note differences where
symlinks are concerned, and the lack of mention of symlinks in `replace` creates a sense
of ambiguity for some users.

This patch clarifies the behavior of `replace` toward symlinks by improving the description
docstring.