Maint: Fix poorly mocked test
Previously, the test was calling into the settings subsystem, which had
the undesirable effect of loading facter, and that generated errors
(though they weren't fatal) due to the expectation on FileTest
interferring with Facter's operatingsystem fact on debian6, at least.

  Could not retrieve operatingsystem: unexpected invocation:
  FileTest.exists?('/etc/debian_version')

This commit changes the test so that we don't calling into the settings
code.