(maint) Fix hiera_include spec tests
Without this patch applied, the hiera_include function spec tests are
not examples of actual behavior.  In the spec tests the stubbed methods
would return an empty array, which would then be forwarded straight away
to the include parser function.  The example did not have the
expectation that this would raise an error.

This behavior is a problem because adding arity checking to the
functions becomes more difficult if theses examples are not
demonstrating exepcted behavior.  Instead, the change to check arity
causes these examples to fail unexpectedly.

This patch fixes the problem by setting the expectations of the actual
behavior of the hiera_include functions more accurately.  An empty array
is no longer returned from the stubbed Hiera behaviors.  This patch
causes us to expect the system to take exception with trying to include
a non-existent class.