Revert "(#15560) Delete user's profile after deleting account"
This reverts commit 5e84c971c4f6142d3e4ed46595e528ea35bac9fb.

This commit has been released in 3.0.0, but should probably not be released in
2.7.20. (#15560) addresses managing of homedirs in Windows - the issue is
deleting a user does not result in the removal of the profile even if
`managehome => true` is set (a feature we claim to support in
http://docs.puppetlabs.com/references/2.7.19/type.html). However, users of
puppet may have come to inadvertently rely on this bug, and expect user
homedirs to remain, even with managehome => true. Suddenly deleting them in
Puppet 2.7.20 may cause serious problems for folks, particularly since deletion
of user data is already a particularly sensitive issue with regulatory/
compliance implications for many organizations.

Conflicts:
	spec/unit/provider/user/windows_adsi_spec.rb