(#17447) Stop puppet before returning in debian init script
Without the retry flag, the stop call in the debian init script isn't
guaranteed to stop the puppet agent or master. This commit adds a 2 step
timeout to the init scripts. A TERM signal is sent first, and after 10 seconds
if the process is still running a KILL signal is sent. The unneccessary sleep
call is also removed from the restart case. This affects both agent and
master init scripts.