(#17445) Update logrotate for debian
Previously logrotate would run the postrotate script n times where n was equal
to the number of logs matching the pattern. This could cause puppet to
misbehave if it received several HUP or USR2 signals at once as noted in the
ticket. This commit adds a sharedscripts line to the logrotate rule, which
tells logrotate to only run the postrotate script once, for all matching log
files found.